Add swipe detection
The problem is that on mobile devices, if you are dragging the page and start the drag on a rome calendar input, the calendar will popup as soon as you finish dragging. I've added a touchstart and touchmove event to figure out if it is a swipe or if touchend should fire the calendar.